# docs/swagger.yaml
basePath: /
definitions:
...
  endpoint.createDiseaseRequest:
    properties:
      description:
        example: A metabolic disease
        type: string
      name:
        example: Diabetes
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Disease:
    description: Disease information
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        example: A metabolic disease
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Diabetes
        type: string
      updatedAt:
        type: string
    type: object
...
  util.APIResponse:
    properties:
      data: {}
      error:
        type: string
      msg:
        type: string
      success:
        type: boolean
    type: object
host: localhost:19091
info:
  contact:
    email: support@ariebrainware.com
    name: Arie Brainware
  description: |-
    REST API for managing patient data, treatments, and therapy sessions
    This API provides endpoints for patient management, disease tracking, treatment records, and therapist management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LTT Backend API
  version: "1.0"
paths:
  /disease:
    get:
      consumes:
      - application/json
      description: Get a paginated list of diseases
      parameters:
      - default: 10
        description: Limit number of results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Diseases retrieved
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Disease'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: List all diseases
      tags:
      - Disease
    post:
      consumes:
      - application/json
      description: Add a new disease to the system
      parameters:
      - description: Disease information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/endpoint.createDiseaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Disease created
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Disease'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Create a new disease
      tags:
      - Disease
  /disease/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a disease by ID
      parameters:
      - description: Disease ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Disease deleted
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Disease not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Delete a disease
      tags:
      - Disease
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific disease
      parameters:
      - description: Disease ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Disease retrieved
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Disease'
              type: object
        "400":
          description: Disease not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Get disease information
      tags:
      - Disease
    patch:
      consumes:
      - application/json
      description: Update an existing disease's information
      parameters:
      - description: Disease ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated disease information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/endpoint.createDiseaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Disease updated
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Disease'
              type: object
        "400":
          description: Invalid request or disease not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Update disease information
      tags:
      - Disease
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SessionToken:
    description: Session token for authenticated requests
    in: header
    name: session-token
    type: apiKey
swagger: "2.0"